Added Competition Version
Messed up last commit, will rectify that here I guess.

Last commit was about the PID Theory lessons for the wiki. They
serve to teach about the theory of PID and how each component
functions to better understand how the control system functions.

This commit is for the competition version folder, whose purpose
is for competition season since the current style of wiki entries
moreso teach from the ground up how the robot works, but aren't
useful as references during competition season to a great degree.
The competition version is going to be loose collections of code
snippets and documentation for certain topics for easy reference.